### 2025-04-06 #8004 @editorialbot
[PRE REVIEW]: p2smi: A Python Toolkit for Peptide FASTA-to-SMILES Conversion and Molecular Property Analysis
**Submitting author:** @AaronFeller (<a href="http://orcid.org/0000-0002-4476-1026">Aaron Feller</a>)
**Repository:** https://github.com/AaronFeller/p2smi
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.1
