### 2025-04-14 167b86d
Update README - Remove paper
