### 2025-05-20 #8004 @lucydot
@himaghna @zhanzmr @gmrandazzo @anandojha - hello 👋  - are you available to review this submission to JOSS?
