### 2025-05-20 #8004 @himaghna
Sure!

Himaghna Bhattacharjee
Machine Learning Research Scientist, Meta Platforms Inc.

*“**You have to want to falsify the model. If you love somebody, set them
free.”* *- Gelman*

On Tue, May 20, 2025 at 7:59 AM Lucy Whalley ***@***.***>
wrote:

> *lucydot* left a comment (openjournals/joss-reviews#8004)
> <https://github.com/openjournals/joss-reviews/issues/8004#issuecomment-2894747560>
>
> @himaghna <https://github.com/himaghna> @zhanzmr @gmrandazzo
> <https://github.com/gmrandazzo> @anandojha <https://github.com/anandojha>
> - hello 👋 - are you available to review this submission to JOSS?
>
