### 2025-05-23 #8004 @lucydot
@gmrandazzo @anandojha - hello 👋 - are you available to review this submission to JOSS?
