### 2025-08-05 #8319 @lucydot
Hi @himaghna and @gmrandazzo - are you still able to review this JOSS submission? The thread has been quiet since May. If I don't hear anything in the next week I'll start looking for other reviewers. If you are still able to review it would be much appreciated.
