### 2025-08-13 #8319 @AaronFeller
Hi @lucydot

Just wanted to check in on the progress of finding new reviewers. I appreciate your support and work on continuing to support open scholarship and scientific publication.

If you'd like I could take another look at the JOSS reviewer list and suggest new reviewers.

Warmly,
Aaron
