### 2025-09-04 #8319 @lucydot
Thank you for your comments @gmrandazzo  - a note regarding installation, also for @AaronFeller. We don't require any project to use a single tool over another e.g. poetry over pip, unless it is required for the software to function or meet the stated use cases.
