### 2025-09-18 #8319 @AnjaConev
Hi all, apologies for the delay - I just got started with the review here.
Here are a few comments for now @AaronFeller I will test out the software tomorrow, so for now my comments are mostly related to the paper text and content and not so much on the tool itself.

1. Statement of need
Here I very much agree with @gmrandazzo - it would be good to mention BioPython and the functionalities of it and highlight the noncanonical amino acids!
Also, you mention: ... "however, many face limitations such as proprietary licensing and lack of _specific functionalities_ for drug-like peptides. " Can you describe a few of those _specific functionalities_ needed for drug-like peptides, just to give us a few concrete examples. I guess you talk more about them in the 'Features' section, but it might be good to explain one or two here too.

2. Data sharing
You mentioned you used this to generate a dataset of 10M peptides for your previous paper. Is this data available somewhere? If yes, it would be nice to give a reference to it either somewhere in the github repo or even in the paper under data availability section.

3. Documentation
Any chance for a more detailed documentation with something like reachthedocs - this is not too necessary as your tool is clear and simple and the github readme might be enough for now, but it would be a welcome addition!
