### 2025-09-18 #8319 @AaronFeller
@AnjaConev and @gmrandazzo, thank you for the reviews. All of your comments are very helpful and point out some strong value adds to the current state of the manuscript and respository.

@lucydot Yes, I've read the comments from both reviewers and think their points are very salient. I'll be working through cleaning up the repository, documentation, and adding some functionality, as well as making suggested edits to the manuscript (which I see as all well warranted).

Also, this open style of review is new to me and I quite like it!

--

@lucydot is there a timeline that I should adhere to for revisions?
