### 2025-10-01 e001750
Revisions for JOSS
