+++ manuscript/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'p2smi: A Python Toolkit for Peptide FASTA-to-SMILES Conversion and Molecular Property Analysis'
+title: 'p2smi: A toolkit enabling SMILES generation and property analysis for noncanonical and cyclized peptides'
 tags:
   - Python
   - cheminformatics
@@ -27,43 +27,63 @@ archive_doi: 10.5281/zenodo.15059960
 ---
 
 # Summary
+Converting peptide sequences into useful representations for downstream analysis is a common step in computational modeling and cheminformatics. Furthermore, peptide drugs (e.g. Semaglutide, Degarelix) include chemistries beyond natural amino acids and standard backbone structure. Common modifications used include noncanonical amino acids, alternate stereochemistry (D- vs L- amino acids), modified chemistrie such as N-methylation and PEGylation. <!-- Of the available chemoinformatics toolkits, none are able to convert these drug-like peptide from an amino acid representation to the chemical string nomenclature SMILES, a step necessary when applying language modeling in cheminformatics. -->Here we present p2smi, a Python toolkit with CLI, designed to facilitate the generation of drug-like peptides and their conversion into chemical SMILES strings. By supporting peptide cyclizations, unnatural amino acid incorporation, and chemical modifications, p2smi enables generation of accurate SMILES strings for drug-like peptides, providing a missing link for computational modeling and cheminformatics analyses. The toolkit offers functionalities for chemical modification of peptides and for calculation of molecular properties such as hydrophobicity, topological polar surface area, molecular weight, and adherence to Lipinski’s rules for drug-likeness.
 
-Converting peptide sequences into useful representations for downstream analysis is a common step in computational modeling and cheminformatics. Furthermore, peptide drugs (e.g. Semaglutide, Degarelix) often take advantage of the diverse chemistries found in noncanonical amino acids (NCAAs), altered stereochemistry, and backbone modifications. Despite there being several chemoinformatics toolkits, none are tailored to the task of converting a modified peptide from an amino acid representation to the chemical string nomenclature Simplified Molecular-Input Line-Entry System (SMILES), often used in chemical modeling. Here we present p2smi, a Python toolkit with CLI, designed to facilitate the conversion of peptide sequences into chemical SMILES strings. By supporting both cyclic and linear peptides, including those with NCAAs, p2smi enables researchers to generate accurate SMILES strings for drug-like peptides, reducing the overhead for computational modeling and cheminformatics analyses. The toolkit also offers functionalities for chemical modification, synthesis feasibility evaluation, and calculation of molecular properties such as hydrophobicity, topological polar surface area, molecular weight, and adherence to Lipinski’s rules for drug-likeness.
 
-
-# Statement of need 
-
[…]
