+++ manuscript/paper.bib
@@ -86,7 +86,38 @@
 @software{p2smi,
   author={Aaron L. Feller},
   title={p2smi: Peptide FASTA-to-SMILES Conversion and Molecular Property Tools},
-  version={1.0.0},
+  version={1.1.0},
   date={2025-03-21},
   url={https://github.com/AaronFeller/p2smi},
+}
+
+@article{cock2009biopython,
+  title={Biopython: freely available Python tools for computational molecular biology and bioinformatics},
+  author={Cock, Peter JA and Antao, Tiago and Chang, Jeffrey T and Chapman, Brad A and Cox, Cymon J and Dalke, Andrew and Friedberg, Iddo and Hamelryck, Thomas and Kauff, Frank and Wilczynski, Bartek and de Hoon, MJL},
+  journal={Bioinformatics},
+  volume={25},
+  number={11},
+  pages={1422},
+  year={2009}
+}
+
+@article{tien2013peptidebuilder,
+  title={PeptideBuilder: A simple Python library to generate model peptides},
+  author={Tien, Matthew Z and Sydykova, Dariya K and Meyer, Austin G and Wilke, Claus O},
+  journal={PeerJ},
+  volume={1},
+  pages={e80},
+  year={2013},
+  publisher={PeerJ Inc.}
+}
+
+@article{yang2025cyclicpeptide,
+  title={cyclicpeptide: a Python package for cyclic peptide drug design},
+  author={Yang, Liu and Cao, Suqi and Liu, Lei and Zhu, Ruixin and Wu, Dingfeng},
+  journal={Briefings in Bioinformatics},
+  volume={26},
+  number={1},
+  pages={bbae714},
+  year={2025},
+  publisher={Oxford University Press}
 }
