+++ README.md
@@ -1,18 +1,20 @@
-# p2smi: Peptide FASTA-to-SMILES Conversion and Molecular Property Tools
+# p2smi: Generation and Analysis of Drug-like Peptide SMILES Strings
 
-**p2smi** is a Python package for generating and modifying peptide SMILES strings from FASTA input and computing molecular properties. It supports cyclic and linear peptides, noncanonical amino acids, and common chemical modifications (e.g., N-methylation, PEGylation).
+**p2smi** is a Python package for generating peptide sequences, converting them to SMILES (including cyclizations and noncanonical amino acids), and evaluating their molecular properties. It supports modification (N-methylation, PEGylation), synthesis feasibility checks, and output in a new peptide/SMILES `.p2smi` format.
 
-This package was released in its current form to support work on the **PeptideCLM** model, described in our [Publication](https://pubs.acs.org/doi/10.1021/acs.jcim.4c01441).
+This toolkit was developed in support of [PeptideCLM](https://pubs.acs.org/doi/10.1021/acs.jcim.4c01441), a smiles-based language model for modified peptides.
 
-> **If you use this tool, please cite the PeptideCLM paper.** A publication for this specific toolkit is forthcoming.
+**Please cite our manuscript if you use this tool.**  
+[arXiv preprint](https://arxiv.org/abs/2505.00719)
 
-## Manuscript
-- [View PDF](manuscript/paper.pdf)
-- [View markdown source](manuscript/paper.md)
 
-## Directory
+## 📄 Manuscript
+- [PDF](manuscript/paper.pdf) | [Markdown](manuscript/paper.md)
 
-- [Features](#features)
+
+## 📁 Directory
+
+- [Features](#Features)
 - [Installation](#installation)
 - [Command-Line Tools](#command-line-tools)
 - [Example Usage](#example-usage)
@@ -21,97 +23,120 @@ This package was released in its current form to support work on the **PeptideCL
 - [License](#license)
 - [Citation](#citation)
 
+
 ## Features
+
+- Generate random peptide sequences (with NCAAs, D-stereochemistry, and cyclization)
 - Convert peptide FASTA files into valid SMILES strings
-- Automatically handle peptide cyclizations (disulfide, head-to-tail, side-chain to N-term, side-chain to C-term, side-chain to side-chain)
-- Modify peptide SMILES with customizable N-methylation and PEGylation
-- Evaluate synthesis feasibility with defined synthesis rules
-- Compute molecular properties: logP, TPSA, molecular formula, and Lipinski rule evaluation
+- Support five cyclization types: disulfide, head-to-tail, sidechain-to-sidechain, sidechain-to-N-term, sidechain-to-C-term
+- Modify SMILES with user-defined N-methylation and PEGylation rates
+- Evaluate synthetic feasibility based on common failure motifs
+- Compute molecular properties (MW, logP, TPSA, Lipinski, etc.)
+
 
 ## Installation
+
+Install from PyPI:
+
 ```bash
 pip install p2smi
 ```
-For development:
+
+For local development:
+
 ```bash
 git clone https://github.com/AaronFeller/p2smi.git
 cd p2smi
 pip install -e .[dev]
 ```
 
+
 ## Command-Line Tools
 
-| Command               | Description                                                     |
-|-----------------------|-----------------------------------------------------------------|
-| `generate-peptides`  | Generate random peptide sequences based on user-defined constraints and modifications |
-| `fasta2smi`          | Convert a FASTA file of peptide sequences into SMILES format    |
[…]
