### 2025-10-01 #8319 @AaronFeller
Thank you @lucydot for the organization, and thank you to the reviewers for their responses! I've pushed edits to the manuscript, codebase, and documentation. PyPI has the updated codebase as well. I've attached responses per comment below.

## Response to comments from @gmrandazzo:

> 1. Statement of needs... I agree, Oeye, Chemaxon are commercial.. I agree that rdkit does not provide some functionalities for peptides, etc. However, there is another solution you are not mentioning: biopython using biopython, and with this code, I can generate random peptides using natural amino acids. […example code…] However, your code's strength is working with unnatural amino acids, which I greatly like! So, I suggest changing the strength of this paper and tool to favour unnatural amino acids. Be more specific! This should apply to the entire manuscript. Rework accordingly to accommodate unnatural amino acids.

**Response: Thank you for the comments! I want to first note that the tool PeptideBuilder came out of the lab I am currently working in! However, p2smi is (as you note) is targeted to drug-like or noncanonical peptides, this includes unnatural amino acids, chemical modifications, and cyclizations, all wrapped in one package. I have updated the manuscript to further highlight the differences between p2smi and other available tools (including PeptideBuilder).**

> 2. Feature section: Describe how the algorithm works for each of the tools you have implemented. Short description. The code is available for details. On synthetic feasibility, I would add more details there since it is a hot topic. How do you evaluate that? How does it work? Ok, the code is there, but it is really important to have more information about it.

**Response: I have added a longer description for each function in the manuscript. The github readme has also been updated to include further details.**

> 3. I do not see any comparisons and metrics against other tools, for example, against BioPython. Can you run that using common amino acids? You should compare similar to it or better. If better, explain how/why better. You want to generate large-scale random peptides? Please also state the throughput of this software.

**Response: Thank you for the suggestion. While I agree that benchmarking against other tools (like BioPython) could be informative, I would note that p2smi serves a distinct purpose: it’s specifically designed to generate and manipulate large-scale libraries of noncanonical and cyclized peptides—use cases that fall outside the scope of standard peptide tools like BioPython, which primarily handle natural amino acids.
As such, direct comparison on basic tasks (e.g., sequence-to-FASTA for canonical peptides) would not fully capture the utility or motivation behind p2smi. The throughput of the tool depends on the specific operation, but generating thousands of sequences and converting them to SMILES (even with NCAA and complex cyclizations) typically takes seconds to a few minutes on a laptop (and could be multi-threaded with a simple bash script). RDKit-based steps (like SMILES modification or property calculation) remain the primary bottleneck, but these are consistent with other cheminformatics tools.**

> 4. Some tools are not working, and examples are needed. For example, synthesis-check output.smi does not work for me.

**Response: Synthesis-check only works on canonical amino acids, this is due to the underlying libraries not functioning with unnatural amino acids. Noncanonical AA analysis would be great to incorporate in the future, but that requires more than is in scope for this manuscript and package.**

> 5. Fasta2smi, what is the reason for this? Can I use OpenBabel? I know that this is not possible, especially if you use unnatural amino acids. Please comment about that.

**Response: These comments have been expanded to note that noncanonical amino acids are not available for other packages.**

> 6. Installation should be done using poetry and not pip

**Response: As noted by the editors, restrictions on package managers is not necessary for JOSS submissions. I will note, I use uv which works quite well with PyPI stored packages.**

## Response to comments from @AnjaConev:
> 1. Here I very much agree with @gmrandazzo - it would be good to mention BioPython and the functionalities of it and highlight the noncanonical amino acids! Also, you mention: ... "however, many face limitations such as proprietary licensing and lack of specific functionalities for drug-like peptides. " Can you describe a few of those specific functionalities needed for drug-like peptides, just to give us a few concrete examples. I guess you talk more about them in the 'Features' section, but it might be good to explain one or two here too.

**Response: Thank you for the very helpful review! The manuscript now addresses the lack of functionalities that we address with p2smi.**

> 2. You mentioned you used this to generate a dataset of 10M peptides for your previous paper. Is this data available somewhere? If yes, it would be nice to give a reference to it either somewhere in the github repo or even in the paper under data availability section.

**Response: A data availability section and a link to this dataset has been added to the manuscript.**

> 3. Any chance for a more detailed documentation with something like reachthedocs - this is not too necessary as your tool is clear and simple and the github readme might be enough for now, but it would be a welcome addition!

**Response: I hope that this tool grows to require a read-the-docs in the future, though at this time to cover the utility of these five functions, I have expanded the descriptions for each function in the github readme and in the manuscript to clear up what they do and how they work.**

Errors:
> 4. When I run smiles-props --help I get the following error: [error]

**Response: This has been corrected. smiles-props --help now works correctly. The issue had been a lack of a main() function.**

> 5. When I run the following: [modify-smiles …] on a small input smiles for alanine I get an error in the modifier.smi file: [Malformed line] [Skipped malformed line]

**Response: The modify-smiles previously required input in a p2smi format. This format is generated during the peptide fasta2smi function and can be passed forward to the modify-smiles function. I have added functionality for standard smiles strings in a csv-style file with one SMILES string per line.**

> 6. When I run synthesis-check -i peptides.fasta (with the list of random peptides) I get the following errors: FAIL: ['Parsing error: not enough values to unpack (expected 2, got 1)']

**Response: This function should now work correctly, but only on peptides with natural amino acids. The underlying libraries are not built to work with unnatural amino acids. In the future, I would like to build out synthesis rules and strategies for determining feasibility for unnatural amino acids, though that is beyond the scope of this work.**

**The error for unnatural amino acids will now read:
>seq_1
HICĪCWATVQ -> FAIL: ['Parsing error: Line contains unnatural amino acids']**
