### 2025-10-20 a04f596
Add user constraints for fasta2smi
