+++ README.md
@@ -1,27 +1,12 @@
+![p2smi banner](./p2smi.png)
+
 # p2smi: Generation and Analysis of Drug-like Peptide SMILES Strings
 
-**p2smi** is a Python package for generating peptide sequences, converting them to SMILES (including cyclizations and noncanonical amino acids), and evaluating their molecular properties. It supports modification (N-methylation, PEGylation), synthesis feasibility checks, and output in a new peptide/SMILES `.p2smi` format.
+**p2smi** is a Python toolkit for peptide design and analysis. 
 
-This toolkit was developed in support of [PeptideCLM](https://pubs.acs.org/doi/10.1021/acs.jcim.4c01441), a smiles-based language model for modified peptides.
+It enables generation of peptide sequences, conversion to SMILES representations—including support for cyclic and noncanonical amino acids—and evaluation of molecular properties. The package also provides utilities for structural modification (e.g., N-methylation, PEGylation), synthesis feasibility assessment, and output in a dedicated .p2smi format that links peptide sequences to their corresponding SMILES.
 
-**Please cite our manuscript if you use this tool.**  
-[arXiv preprint](https://arxiv.org/abs/2505.00719)
-
-
-## 📄 Manuscript
-- [PDF](manuscript/paper.pdf) | [Markdown](manuscript/paper.md)
-
-
-## 📁 Directory
-
-- [Features](#Features)
-- [Installation](#installation)
-- [Command-Line Tools](#command-line-tools)
-- [Example Usage](#example-usage)
-- [Future Work](#future-work)
-- [Contributing](#for-contributors)
-- [License](#license)
-- [Citation](#citation)
+Developed in support of [PeptideCLM](https://pubs.acs.org/doi/10.1021/acs.jcim.4c01441), a SMILES-based language model for modified peptides, p2smi provides an extensible foundation for computational peptide chemistry and machine-learning-driven molecular design.
 
 
 ## Features
@@ -32,6 +17,35 @@ This toolkit was developed in support of [PeptideCLM](https://pubs.acs.org/doi/1
 - Modify SMILES with user-defined N-methylation and PEGylation rates
 - Evaluate synthetic feasibility based on common failure motifs
 - Compute molecular properties (MW, logP, TPSA, Lipinski, etc.)
+
+## Updates
+- Version 1.1.1 - Added functionality to allow for user-defined cyclizing residue constraints
+- Version 1.1.0 - Updated codebase, documentation, fixed bugs -- for JOSS review
+- Version 1.0.0 - First release for JOSS submission
+
+
+## Citation
+
+If you use this tool, please cite:
+
+*p2smi: A Python Toolkit for Peptide FASTA-to-SMILES Conversion and Molecular Property Analysis*.  
+Feller, A. L. and Wilke, C. O. (2025).  
+[arXiv](https://arxiv.org/abs/2505.00719)
+
+A JOSS publication for this package is in review.
+
+## Manuscript
+- [PDF](manuscript/paper.pdf) | [Markdown](manuscript/paper.md)
+
+
+## Directory
+
+- [Installation](#installation)
+- [Command-Line Tools](#command-line-tools)
+- [Example Usage](#example-usage)
+- [Future Work](#future-work)
+- [Contributing](#for-contributors)
+- [License](#license)
 
 
 ## Installation
@@ -56,7 +70,7 @@ pip install -e .[dev]
 | Command             | Description |
 |-------------------------|-------------------------------------------------------------------------------------------------------------------------|
[…]
