### 2025-10-20 31314c5
Add user constraints for fasta2smi (#8)
* Add user constraints for fasta2smi
* linting
