+++ p2smi/fasta2smi.py
@@ -11,6 +11,9 @@ Uses p2smi.utilities.smilesgen.
 import argparse
 import p2smi.utilities.smilesgen as smilesgen
 
+all_aminos = smilesgen.all_aminos
+LETTER2NAME = smilesgen.LETTER2NAME
+
 
 class InvalidConstraintError(Exception):
     # Custom exception for invalid constraints
@@ -67,14 +70,139 @@ def constraint_resolver(sequence, constraint):
         raise InvalidConstraintError(f"{sequence} has invalid constraint {constraint}")
 
 
+def has_capability(aa, key):
+    """
+    Return True if the amino acid supports the given capability key.
+    Works whether the value is a bool, 'False' string, or a SMILES fragment.
+    """
+    val = aa.get(key, False)
+    if val is None:
+        return False
+    if isinstance(val, str):
+        # strip whitespace and quotes, and check for the literal word False
+        val_stripped = val.strip().lower()
+        if val_stripped in {"", "false", "none"}:
+            return False
+        return True
+    return bool(val)
+
+
+def validate_constraint_pattern(peptideseq, pattern):
+    """
+    Validate that a user-supplied constraint pattern matches both
+    the peptide chemistry (using all_aminos) and the positional mask.
+
+    Uses only 'N' for nucleophilic sidechains (no 'E').
+    Returns (is_valid, message) instead of raising exceptions.
+    """
+
+    seq = "".join(peptideseq) if not isinstance(peptideseq, str) else peptideseq
+    tag = pattern[:2].upper()
+    mask = pattern[2:]
+
+    if not pattern:
+        return True, "No constraint pattern provided."
+
+    if len(mask) != len(seq) and len(mask) > 0:
+        return (
+            False,
+            f"Mask length ({len(mask)}) does not match sequence length ({len(seq)}).",
+        )
+
+    try:
+        residues = [all_aminos[LETTER2NAME[r]] for r in seq]
+    except KeyError as e:
+        return False, f"Undefined residue {e.args[0]} in sequence."
+
+    # -----------------------------
+    # Validate per-position codes
+    # -----------------------------
+    for i, code in enumerate(mask):
+        if code == "X":
+            continue
+        aa = residues[i]
+        if code == "C" and not has_capability(aa, "disulphide"):
+            return False, f"Position {i}: '{seq[i]}' cannot form disulfide ('C')."
+        if code == "N" and not has_capability(aa, "nterm"):
+            return False, f"Position {i}: '{seq[i]}' lacks nterm capability ('N')."
+        if code == "Z" and not has_capability(aa, "cterm"):
+            return False, f"Position {i}: '{seq[i]}' lacks cterm capability ('Z')."
+
+    # -----------------------------
+    # Constraint-type validation
+    # -----------------------------
+    if tag == "SS":
+        if sum(has_capability(aa, "disulphide") for aa in residues) < 2:
+            return False, "SS constraint requires ≥2 disulphide-capable residues."
+        return True, "Valid SS constraint."
+
+    elif tag == "HT":
+        if len(residues) < 2:
+            return False, "HT requires at least two residues."
+        return True, "Valid HT constraint."
+
+    elif tag == "SC":
+        codes = set(mask)
+        if "N" in codes and "Z" in codes:
+            subtype = "SCSC"  # sidechain–sidechain
+        elif "N" in codes:
+            subtype = "SCNT"  # sidechain–N-terminus (now covers previous 'E')
+        elif "Z" in codes:
+            subtype = "SCCT"  # sidechain–C-terminus
+        else:
+            return False, f"Unrecognized SC pattern: {pattern}"
+
+        # subtype-specific residue chemistry
+        if subtype == "SCSC":
[…]
