+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "p2smi"
-version = "1.1.0"
+version = "1.1.1"
 description = "A toolkit enabling SMILES generation and property analysis for noncanonical and cyclized peptides."
 readme = "README.md"
 license = { text = "MIT" }
