### 2025-11-10 #8319 @AaronFeller
Thanks all, really great process. @lucydot, I've answered the questions below.

> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
> * Double check authors and affiliations (including ORCIDs)
* Authors and ORCIDs are correct.
> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
* The latest changes are released. Version number is 1.1.1: [PyPI](https://pypi.org/project/p2smi/1.1.1/) and [Github](https://github.com/AaronFeller/p2smi/releases/tag/v1.1.1)
> * Archive the release on Zenodo/figshare/etc and post the DOI here.
* Newest pull now on Zenodo. DOI: [10.5281/zenodo.15059959](https://doi.org/10.5281/zenodo.15059959)
> * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
* Yes, updated contributors with ORCIDs.
> * Make sure that the license listed for the archive is the same as the software license.
* Yes, MIT license on both Zenodo and Github.
